+++ examples/arg_print.py
@@ -63,5 +63,6 @@ class ArgumentAndOptionPrinter(cmd2.Cmd):
 
 
 if __name__ == '__main__':
+    import sys
     app = ArgumentAndOptionPrinter()
-    app.cmdloop()
+    sys.exit(app.cmdloop())