+++ examples/hello_cmd2.py
@@ -6,6 +6,7 @@ This is intended to be a completely bare-bones cmd2 application suitable for rap
 from cmd2 import cmd2
 
 if __name__ == '__main__':
+    import sys
     # If run as the main application, simply start a bare-bones cmd2 application with only built-in functionality.
 
     # Set "use_ipython" to True to include the ipy command if IPython is installed, which supports advanced interactive
@@ -13,4 +14,4 @@ if __name__ == '__main__':
     app = cmd2.Cmd(use_ipython=True, persistent_history_file='cmd2_history.txt')
     app.locals_in_py = True     # Enable access to "self" within the py command
     app.debug = True            # Show traceback if/when an exception occurs
-    app.cmdloop()
+    sys.exit(app.cmdloop())