+++ examples/remove_unused.py
@@ -26,5 +26,6 @@ class RemoveUnusedBuiltinCommands(cmd2.Cmd):
 
 
 if __name__ == '__main__':
+    import sys
     app = RemoveUnusedBuiltinCommands()
-    app.cmdloop()
+    sys.exit(app.cmdloop())